requirements.txt
flask>=3.1.1
gunicorn>=23.0.0
langchain>=0.3.25
langchain-community>=0.3.25
langchain-core>=0.3.65
langchain-milvus>=0.2.0
langchain-ollama>=0.3.3
langchain-openai>=0.3.23
pymilvus>=2.5.11
pypdf>=5.6.0
python-dotenv

PyMuPDF==1.25.0
Pillow==10.4.0
pytesseract==0.3.10
requests==2.32.3 
flasgger